Step - 6 Collecting Collectable Collectables
In this step, I added scripting for the Pick Up objects to react to and
disappear on player collisions. This was done by using the CompareTag
and SetActive functions. Additionally, I learned about how unity is
constantly calculating the positions of static objects, so I gave the
Pick Up objects RigidBody physics instead...