Step 7 - Counters
Added scripting to the player control object that allowed for a counter
to be incremented each time the collectible triggers were activated.
This allows for the player to know how many collectables that have
collected in the top left corner of the screen.

Once they've collected all of them, text saying "You Win!!!" pops up
near the center of the screen.